fix(nuvai-mkl): address PR #18 review findings
- vsl: surface non-finite uniform spans as errors instead of panicking
- aarch64-sys: declare vDSP_DFT_RealtoComplex as u8 (_Bool), not c_int
- fft: select the vDSP family deterministically and report a null setup
  on a supported length honestly instead of probing (and conflating OOM)
- pardiso: pin !Send+!Sync across backends, add reset(), drop the
  factorization-cache .expect
- build: enforce macOS 12.0 deployment target for interleaved vDSP
- tests: cover the FFT length-2 split fallback and non-finite VSL spans

+++ crates/nuvai-mkl-sys/src/aarch64.rs
@@ -57,10 +57,10 @@ pub type vDSP_DFT_Interleaved_Setup = *mut c_void;
 pub type vDSP_DFT_Interleaved_SetupD = *mut c_void;
 
 /// Real-to-complex flag for the interleaved DFT (`vDSP_ENUM(bool,
-/// vDSP_DFT_RealtoComplex)`). Declared `c_int` for ABI uniformity with the
-/// other vDSP enum args; only `0`/`1` is ever passed, so the callee's `_Bool`
-/// read is correct.
-pub type vDSP_DFT_RealtoComplex = c_int;
+/// vDSP_DFT_RealtoComplex)`). The C enum's underlying type is `_Bool` (1 byte),
+/// so it is declared `u8` — not `c_int` (4 bytes) — to match the callee's ABI
+/// register width on ARM64. Only `0`/`1` is ever passed.
+pub type vDSP_DFT_RealtoComplex = u8;
 pub const vDSP_DFT_Interleaved_ComplextoComplex: vDSP_DFT_RealtoComplex = 0;
 pub const vDSP_DFT_Interleaved_RealtoComplex: vDSP_DFT_RealtoComplex = 1;
